penalty: don't use [] as default argument
The default value for an implicit argument is allocated
when the function is created, not when it's called.

Each time the function is called, the value is still hanging
around from the last call. This gets even weirder if you throw
threads into the mix. If two different threads are executing
the function at the same time, and one of them changes a default
argument, they both will see the change.

Signed-off-by: Lucian Adrian Grijincu <lucian.grijincu@gmail.com>