executor: support for .vmx from user submissions
Also added a few options to configure some timeouts:
* timeout for powerOn() - which is now done async on another thread
* delay between when a machine is powered on and when testing begins.

The user can select whether he wants to revert to a snapshot or just
boot the machine.

Also, because vmware asks if a new machine it encounters was moved or
copied, we need to interact with it (bin/vmchecker-vm-message-handler)
and say it was only moved.

Signed-off-by: Lucian Adrian Grijincu <lucian.grijincu@gmail.com>