submissions: add methods to parse EvaluationQueueingTime
The time a submission is sent to evaluation may differ from the time
it's uploaded.

In MD5Submissions, the user uploads an MD5 on one machine and a
virtual machine somewhere else. The moment the MD5's is stored is
recorded as the 'UploadTime', but the submission will be evaluated at
a later time, when the virtual machine is uploaded somewhere else.

We're going to keep these as separate fields in the submission config
file.

Signed-off-by: Lucian Adrian Grijincu <lucian.grijincu@gmail.com>